+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 - 2019-11-26: Consistently refer to MocoTrajectories as "trajectory" instead of
               "iterate".
 
+- 2019-11-22: Introduce TabOpUseAbsoluteStateNames to convert column labels
+              from IK solutions pre-4.0 states files to use new-style column
+              labels.
+
 - 2019-11-09: Improve the reliability of building the Ipopt dependency by using
               more stable servers for downloading Metis and MUMPS.